Move DefaultHttpClient and defaultLogger out of request-handler.ts to remain congruent with their optionality
