+++ src/core/defaults/index.ts
@@ -1,0 +1,2 @@
+export { DefaultHttpClient } from './http-client.ts';
+export { defaultLogger } from './logger.ts';
