+++ src/core/defaults/logger.ts
@@ -1,0 +1,8 @@
+import { Logger } from '../types.ts';
+
+export const defaultLogger: Logger = {
+  debug: console.debug.bind(console),
+  info: console.info.bind(console),
+  warn: console.warn.bind(console),
+  error: console.error.bind(console),
+};
