+++ src/index.ts
@@ -1,9 +1,5 @@
-import {
-  DefaultHttpClient,
-  defaultLogger,
-  RequestBuilder,
-  RequestHandler,
-} from './core/request-handler.ts';
+import { RequestBuilder, RequestHandler } from './core/request-handler.ts';
+import { DefaultHttpClient, defaultLogger } from './core/defaults/index.ts';
 import { XmApiOptions } from './core/types.ts';
 import { GroupsEndpoint } from './endpoints/groups/index.ts';
